backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.0
matplotlib==3.8.2
numpy==1.26.4
